610-Assistant/backend/requirements.txt
django
djangorestframework
gunicorn
dj-database-url
psycopg2-binary
whitenoise
